packages/qeeg-encoder/pyproject.toml
[project]
name = "qeeg-encoder"
version = "0.1.0"
description = "Brain Twin qEEG modality encoder — foundation + tabular dual path with MAPIE conformal wrapping."
requires-python = ">=3.11"
authors = [{ name = "DeepSynaps Studio" }]
license = { text = "Proprietary" }
dependencies = [
  "aiokafka>=0.11.0",
  "fastavro>=1.9.4",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "numpy>=1.26.0",
  "scipy>=1.13.0",
  "scikit-learn>=1.5.0",
  "mapie>=0.9.0",
  "torch>=2.3.0",
  "mne>=1.7.0",
  "feast>=0.40.0",
  "redis>=5.0.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "structlog>=24.1.0",
  "pyyaml>=6.0.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-pyyaml",
  "hypothesis>=6.100.0",
]

[project.scripts]
qeeg-encoder = "qeeg_encoder.cli:main"
qeeg-encoder-consume = "qeeg_encoder.bus.consumer:run"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "S", "RUF"]
ignore = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --cov=qeeg_encoder --cov-report=term-missing"

